+++ tests/index.js
@@ -2,7 +2,7 @@ require("./lastfm-node-test.js");
 require("./lastfm-recenttracks-parser-test.js");
 require("./lastfm-recenttracks-stream-test.js");
 require("./lastfm-session-test.js");
-require("./lastfm-response-reader-test.js");
 require("./lastfm-info-test.js");
 require("./lastfm-info-track-test.js");
 require("./lastfm-update-test.js");
+require("./lastfm-request-test.js");